Switch on travis's caching features
Activated both `cargo` and arbitrary directory. Strictly,
the definition of the `cargo` cache type just takes the
whole `target` dir from the route, which will happen to
pick up OpenSSL's static libs as well, but since that's
not actually an output of cargo, it seems better to be
explicit. This way we don't have to worry about the case
that travis gets more targetted caching rules.

This is the start of investigating #1027